+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Remove trending tags from the list of participated tags we're pulling toots for because we get them anyways
 * Fix bug where home timeline toot cache would just grow (dedupe on the way in)
 * Add DEFAULT_REQUESTS configuration
+* Rename ScoreInfo type to WeightInfo, rename TheAlgorithm.scorersDict to weightInfo
 
 # v0.34.0
 * Use a value for the user displayed TIME_DECAY param that is 10x the value used in the calculation to make it more user friendly.